+++ ToDosAPI/Models/Dtos/EditTaskDto.cs
@@ -4,7 +4,7 @@ namespace ToDosAPI.Models.Dtos
 {
     public class EditTaskDto
     {
-        [Required] public int TaskId { get; set; } = default!;
+        [Required] public int Id { get; set; } = default!;
         [Required] public string TaskContent { get; set; } = default!;
         [Required] public int Status { get; set; } = default!;
     }