+++ client/changes.md
@@ -11,10 +11,10 @@
 ### 22-11-2023
 
 - Login on enter key press -
-- Add error indicator if login is not successful (ngx-toastr)
+- Add error indicator if login is not successful (ngx-toastr) -
 - Install prettier extension && enable Format on Save from settings -
 - Move images folder path to appsettings -
-- Add a logo to the website, must be png
+- Add a logo to the website, must be png -
 - Create a new component to display every todo
 - Make it highlighted if it's completed, otherwise no, also on click -> complete. on click again -> uncomplete
-- Add logout button
+- Add logout button -